ai-paper-evaluator/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.1.0
google-generativeai==0.8.5
PyPDF2==3.0.1
pdfplumber==0.11.6
Werkzeug==3.1.3
SQLAlchemy==2.0.36